// inetum-realjobs-frontend/src/app/models/functionCategory.enum.ts
export enum FunctionCategory {
  GENERAL_MANAGEMENT = "GENERAL_MANAGEMENT",
  EDUCATION_AND_TRAINING = "EDUCATION_AND_TRAINING",
  LAW = "LAW",
  RESEARCH_AND_DEVELOPMENT = "RESEARCH_AND_DEVELOPMENT",
  HEALTH_MEDICINE_SOCIAL_AND_FITNESS = "HEALTH_MEDICINE_SOCIAL_AND_FITNESS",
  ACCOMMODATION_CATERING_AND_TOURISM = "ACCOMMODATION_CATERING_AND_TOURISM",
  LAND_SEA_ENVIRONMENT_AND_VETERINARY = "LAND_SEA_ENVIRONMENT_AND_VETERINARY",
  PUBLIC_SERVICES = "PUBLIC_SERVICES",
  POLICE_SECURITY_AND_PROTECTIVE = "POLICE_SECURITY_AND_PROTECTIVE",
  MEDIA_AND_INFORMATION = "MEDIA_AND_INFORMATION",
  IT_AND_TELECOMMUNICATIONS = "IT_AND_TELECOMMUNICATIONS",
  SALES_AND_PURCHASING = "SALES_AND_PURCHASING",
  BANKING_FINANCE_SERVICES_AND_INSURANCE = "BANKING_FINANCE_SERVICES_AND_INSURANCE",
  FINANCE_ADMINISTRATION_AND_AUDITING = "FINANCE_ADMINISTRATION_AND_AUDITING",
  HR = "HR",
  MARKETING_AND_ADVERTISING = "MARKETING_AND_ADVERTISING",
  ADMINISTRATIVE_SUPPORT_AND_SECRETARIAL = "ADMINISTRATIVE_SUPPORT_AND_SECRETARIAL",
  TRANSPORT_AND_LOGISTICS = "TRANSPORT_AND_LOGISTICS",
  MANUFACTURING_BUILDING_AND_MANUAL_WORK = "MANUFACTURING_BUILDING_AND_MANUAL_WORK",
  CULTURE_AND_FINE_ARTS = "CULTURE_AND_FINE_ARTS",
  ENGINEERING_AND_TECHNICAL_JOBS = "ENGINEERING_AND_TECHNICAL_JOBS",
  DESIGN_AND_ARCHITECTURE = "DESIGN_AND_ARCHITECTURE"
}

export const FUNCTION_CATEGORY_OPTIONS = [
  {name: "General management", value: FunctionCategory.GENERAL_MANAGEMENT},
  {name: "Education & training", value: FunctionCategory.EDUCATION_AND_TRAINING},
  {name: "Law", value: FunctionCategory.LAW},
  {name: "R&D", value: FunctionCategory.RESEARCH_AND_DEVELOPMENT},
  {name: "Health, medicine, social & fitness", value: FunctionCategory.HEALTH_MEDICINE_SOCIAL_AND_FITNESS},
  {name: "Accommodation, catering & tourism", value: FunctionCategory.ACCOMMODATION_CATERING_AND_TOURISM},
  {name: "Land, sea, environment & veterinary", value: FunctionCategory.LAND_SEA_ENVIRONMENT_AND_VETERINARY},
  {name: "Public services", value: FunctionCategory.PUBLIC_SERVICES},
  {name: "Police, security & protective", value: FunctionCategory.POLICE_SECURITY_AND_PROTECTIVE},
  {name: "Media & information", value: FunctionCategory.MEDIA_AND_INFORMATION},
  {name: "IT & telecommunications", value: FunctionCategory.IT_AND_TELECOMMUNICATIONS},
  {name: "Sales & purchasing", value: FunctionCategory.SALES_AND_PURCHASING},
  {name: "Banking, finance services & insurance", value: FunctionCategory.BANKING_FINANCE_SERVICES_AND_INSURANCE},
  {name: "Finance administration & auditing", value: FunctionCategory.FINANCE_ADMINISTRATION_AND_AUDITING},
  {name: "HR", value: FunctionCategory.HR},
  {name: "Marketing & advertising", value: FunctionCategory.MARKETING_AND_ADVERTISING},
  {name: "Administrative support & secretarial", value: FunctionCategory.ADMINISTRATIVE_SUPPORT_AND_SECRETARIAL},
  {name: "Transport & logistics", value: FunctionCategory.TRANSPORT_AND_LOGISTICS},
  {name: "Manufacturing, building and manual work", value: FunctionCategory.MANUFACTURING_BUILDING_AND_MANUAL_WORK},
  {name: "Culture & fine arts", value: FunctionCategory.CULTURE_AND_FINE_ARTS},
  {name: "Engineering & technical jobs", value: FunctionCategory.ENGINEERING_AND_TECHNICAL_JOBS},
  {name: "Design & architecture", value: FunctionCategory.DESIGN_AND_ARCHITECTURE},
];